Update files to make testing works + create hidden gem interface in a separate ts file
