+++ src/main/webapp/package.json
@@ -20,9 +20,7 @@
     "@angular/platform-browser": "^11.0.9",
     "@angular/platform-browser-dynamic": "^11.0.9",
     "@angular/router": "^11.0.9",
-    "chromium": "^3.0.2",
     "rxjs": "~6.6.0",
-    "rxjs-compat": "^6.6.3",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
@@ -31,7 +29,7 @@
     "@angular/cli": "^11.0.7",
     "@angular/compiler-cli": "^11.0.9",
     "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.19.14",
+    "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
