+++ src/main/webapp/src/app/hidden-gem.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { HiddenGem } from './hidden-gem';
 
 @Injectable({
   providedIn: 'root'
@@ -12,18 +13,4 @@ export class HiddenGemService {
   getAllHiddenGems() : Observable<HiddenGem[]> {
     return this.httpClient.get<HiddenGem[]>('/hiddengems')
   }
-}
-
-export interface HiddenGem {
-  id: number;
-  name: string;
-  business_type: string;
-  address: string;
-  price_level: number;
-  rating: number;
-  photo: string;
-}
-
-interface HiddenGemsResponse {
-  hiddenGems: HiddenGem[];
 }
