+++ src/main/webapp/src/app/hidden-gem.ts
@@ -1,0 +1,9 @@
+export interface HiddenGem {
+  id: number;
+  name: string;
+  business_type: string;
+  address: string;
+  price_level: number;
+  rating: number;
+  photo: string;
+}
