+++ src/main/webapp/src/main.ts
@@ -1,5 +1,3 @@
-import 'zone.js/dist/zone-testing';
-
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
