+++ src/main/webapp/src/polyfills.ts
@@ -56,9 +56,6 @@
  * Zone JS is required by default for Angular itself.
  */
 import 'zone.js/dist/zone';  // Included with Angular CLI.
-import 'zone.js/dist/zone-testing';
-import 'zone.js/dist/proxy';
-import "zone.js/dist/jasmine-patch";
 
 
 /***************************************************************************************************
