+++ src/main/webapp/src/test.ts
@@ -1,12 +1,14 @@
 // This file is required by karma.conf.js and loads recursively all the .spec and framework files
 
 import 'zone.js/dist/zone-testing';
+
 import 'zone.js/dist/proxy';
 import 'zone.js/dist/jasmine-patch';
 import 'zone.js/dist/long-stack-trace-zone';
 import 'zone.js/dist/sync-test';
 import 'zone.js/dist/async-test';
 import 'zone.js/dist/fake-async-test';
+
 import { getTestBed } from '@angular/core/testing';
 import {
   BrowserDynamicTestingModule,
