+++ src/main/webapp/tsconfig.json
@@ -15,7 +15,7 @@
     "moduleResolution": "node",
     "importHelpers": true,
     "target": "es2015",
-    "module": "amd",
+    "module": "es2020",
     "lib": [
       "es2018",
       "dom"
